Gave up on UIText for now, just started making the dialogue manager

+++ Assets/Scripts/CharacterObject.cs
@@ -8,9 +8,9 @@ public class CharacterObject : ScriptableObject
     public string characterName;
     public string characterVoice;
     [Range(0,100)]
-    public float characterVoicePitch;
+    public int characterVoicePitch;
     [Range(0, 100)]
-    public float characterVoiceRange;
+    public int characterVoiceRange;
     [Range(80, 450)]
-    public float characterVoiceRate;
+    public int characterVoiceRate;
 }
